Split out a base class.
